Refactor downstream adapter used for testing.
The MessageDiscardingDownstreamAdapter does not contain any telemetry
specific code.
